Move slot-related components to a separate file
