+++ newdle/client/src/components/TimelineHeader.js
@@ -1,0 +1,29 @@
+import _ from 'lodash';
+import React from 'react';
+import PropTypes from 'prop-types';
+import moment from 'moment';
+import styles from './Timeline.module.scss';
+
+export default function TimelineHeader({hourSeries, hourSpan, hourStep}) {
+  return (
+    <div className={styles['timeline-hours']}>
+      {_.range(0, hourSpan + hourStep, hourStep).map((i, n) => (
+        <div
+          className={styles['timeline-hour']}
+          key={`timeline-label-${i}`}
+          style={{left: `${(i / hourSpan) * 100}%`}}
+        >
+          <span className={styles['timeline-hour-text']}>
+            {moment({hours: hourSeries[n]}).format('k')}
+          </span>
+        </div>
+      ))}
+    </div>
+  );
+}
+
+TimelineHeader.propTypes = {
+  hourSeries: PropTypes.array.isRequired,
+  hourSpan: PropTypes.number.isRequired,
+  hourStep: PropTypes.number.isRequired,
+};
